Fixed new sessions window not opened when receving activity intent from nethunter app.
Found out when nethunter app sending activity intent to nhterm, if nhterm is first opened by that intent sent by nethunter,
then all the next activity intent with the same action and category will not be process, as it's a android mechenism.
see https://stackoverflow.com/questions/3781182/activity-not-started-its-current-task-has-been-brought-to-the-front

The way I found to solve this issue is to added 'android:launchMode="singleTask"' to the corresponding activity.