Fixed Ctrl key not detected by default system keyboard input method.
If user is using default system keyboard input method, the committext() is called,
otherwise, the sendKeyEvent() is called instead. So it needs to be handled different way.

Also I still find no other way to send the meta key combination without using KeyEvent(),
and only one meta key is allowed to be combined at a time for now.